requirements.txt
# Core dependencies
numpy
pandas
matplotlib
seaborn

# Deep learning framework
torch
torchvision
torchinfo

# Image processing
Pillow

# Jupyter environment
jupyter
ipython

# Utility libraries
scikit-learn
tqdm
